+++ engine/seeding_engine/__init__.py
@@ -1,3 +1,3 @@
 """libtorrent seeding engine process."""
 
-__version__ = "1.5.0"
+__version__ = "1.6.0"